shout.m4 has been completely rewritten for libshout 2.
It relies on pkg-config or shout-config (in that order) for compiler/linker
flags. --with-shout-prefix is gone. Also gone is --disable-shouttest. I don't
understand the advantages of that home built script over the standard header/
func checks, so I used those. I guess AC_TRY_RUN isn't invoked, but I feel
(mildly) that the business of autoconf is to ensure the program builds
correctly, not runs correctly. Run errors can usually be fixed after the build
(by tweaking LD_LIBRARY_PATH, ld.so.conf, crle etc) without rebuilding.

I've ported ices 0 to this macro in order to test it, but haven't put the various
fallbacks through the ringer yet. I will before release.

TODO: I don't think this file is being included with make dist. Should it be?
Probably. Should it be installed? Probably, but less so.

svn path=/trunk/libshout/; revision=4939