Fix: Check server's certificate in TLS mode.
This adds support to verify the certificate as provided by the server.
Support is limited to CommonName checks. Other values are ignored.
X509_check_host() would be the right way to go. However this is
only in OpenSSL 1.1.0 wich is by this writing not really widespread.

Todo: One should find out how to test for X509_check_host() in
      configure. If present XXX_HAVE_X509_check_host should be set.
      This const may also be renamed.
Closes: #2152