Cleanup: Renamed files to match what they do
This renames files to codec_* and format_* similar to my
patch to move protocol stuff to proto_*.
This is supposted to improve understanding of the src/ directory.

mp3.c, while taking care of a codec is moved to format_mp3.c
as it is handled like a container by libshout as there is no
container around it.
I think this is why it's 'format' not 'container' anyway.