Fix: Escape mountpoint names.
Prior to this commit mountpoint names were sent to the server as-is
both as part of the SOURCE (or eq.) request as well as part of out-
of-band meta data updates (ICY workaround). This was a source for a
potential problem e.g. if the mountpoint's name contains characters
from the reserved space. This commit fixes this by always doing the
escaping step before sending this to the server. This may only be a
problem in case GET parameters were included in the mountpoint URI.
However this was never part of the concept and was supported by the
main line Icecast2 server. Please test this and feel free to reopen
the ticket in case of any problems. Thank you for reporting and for
your work.

Closes: #2233