Fix: Fixed safechar table to match RFC3986 more closely.
The original table is off-RFC. I have no idea what the intention
of that is and why it is that super restrictive. I fixed this on
the new table for resoruce encodings by adding the safe chars as
per RFC3986 section 2.3 to the list. Those extra chars are: 0x2D
('-'), 0x2E ('.'), 0x5F ('_') as well as 0x7E ('~').

See: #2233